fix Driver authen

+++ src/services/driver/index.js
@@ -10,6 +10,7 @@ const END_POINTS = {
 }
 
 export const addDriverAPI = async (driver) => {
+    const user = JSON.parse(localStorage.getItem('user'));
     const formData = new FormData();
     formData.set("email", driver.email);
     formData.set("code", driver.code);
@@ -21,18 +22,20 @@ export const addDriverAPI = async (driver) => {
     formData.set("personalEmail", driver.personalEmail);
     formData.set("dateOfBirth", driver.dateOfBirth);
     formData.set("avatarFile", driver.avatarFile); // Use 'avatarFile' as the key
-    
-    return await axios.post(`${URL}/${END_POINTS.addDriver}`, 
-    formData, 
-    {
-      headers: {
-        Authorization: `Bearer ${JSON.parse(localStorage.getItem("user")).accessToken}`,
-        "Content-Type": "multipart/form-data", // Set the content type
-      },
-    });
-  };
-  
-  export const updateDriverAPI = async (driver, id) => {
+    if (user && user.accessToken) {
+        return await axios.post(`${URL}/${END_POINTS.addDriver}`,
+            formData,
+            {
+                headers: {
+                    Authorization: `Bearer ${user.accessToken}`,
+                    "Content-Type": "multipart/form-data", // Set the content type
+                },
+            });
+    }
+};
+
+export const updateDriverAPI = async (driver, id) => {
+    const user = JSON.parse(localStorage.getItem('user'));
     const formData = new FormData();
     formData.append("email", driver.email);
     formData.append("code", driver.code);
@@ -43,62 +46,77 @@ export const addDriverAPI = async (driver) => {
     formData.append("phoneNumber", driver.phoneNumber);
     formData.append("personalEmail", driver.personalEmail);
     formData.append("dateOfBirth", driver.dateOfBirth);
-    formData.append("avatarFile", driver.avatarFile); 
-  
+    formData.append("avatarFile", driver.avatarFile);
+    if (user && user.accessToken) {
     return await axios.put(`${URL}/${END_POINTS.updateDriver}/${id}`, formData, {
-      headers: {
-        Authorization: `Bearer ${JSON.parse(localStorage.getItem("user")).accessToken}`,
-        "Content-Type": "multipart/form-data",
-      },
+        headers: {
+            Authorization: `Bearer ${user.accessToken}`,
+            "Content-Type": "multipart/form-data",
+        },
     });
-  };  
+}
+};
 
 export const getSingleDriver = async (id) => {
+    const user = JSON.parse(localStorage.getItem('user'));
+    if (user && user.accessToken) {
     return await axios.get(`${URL}/${END_POINTS.getSingleDriver}/${id}`,
         {
             headers: {
-                Authorization: `Bearer ${JSON.parse(localStorage.getItem('user')).accessToken}`,
+                Authorization: `Bearer ${user.accessToken}`,
             },
         })
+    }
 }
 
 export const getMultiDriversAPI = async (data) => {
+    const user = JSON.parse(localStorage.getItem('user'));
+    if (user && user.accessToken) {
     return await axios.get(`${URL}/${END_POINTS.getMultiDrivers}`, {
         params: {
             ...data
         },
         headers: {
-            Authorization: `Bearer ${JSON.parse(localStorage.getItem('user')).accessToken}`,
+            Authorization: `Bearer ${user.accessToken}`,
         },
     })
 }
+}
 
 export const getAllDrivers = async () => {
+    const user = JSON.parse(localStorage.getItem('user'));
+    if (user && user.accessToken) {
     return await axios.get(`${URL}/${END_POINTS.getMultiDrivers}`, {
         headers: {
-            Authorization: `Bearer ${JSON.parse(localStorage.getItem('user')).accessToken}`,
+            Authorization: `Bearer ${user.accessToken}`,
         },
     })
+}
 }
 
 
 export const deleteDriverAPI = async (driverId) => {
-    console.log(`[FROM API]: driverId ${driverId}`)
+    const user = JSON.parse(localStorage.getItem('user'));
+    if (user && user.accessToken) {
     return await axios.delete(`${URL}/${END_POINTS.deleteDriver}/${driverId}`,
         {
             headers: {
-                Authorization: `Bearer ${JSON.parse(localStorage.getItem('user')).accessToken}`,
+                Authorization: `Bearer ${user.accessToken}`,
             },
         });
+    }
 }
 
 export const toggleStatusAPI = async (driverId, status) => {
+    const user = JSON.parse(localStorage.getItem('user'));
+    if (user && user.accessToken) {
     return await axios.patch(`${URL}/${END_POINTS.enableDriver}/${driverId}`,
-    status,
-    {
-        headers: {
-            Authorization: `Bearer ${JSON.parse(localStorage.getItem('user')).accessToken}`,
-            "Content-Type": "application/json"
-        },
-    })
+        status,
+        {
+            headers: {
+                Authorization: `Bearer ${user.accessToken}`,
+                "Content-Type": "application/json"
+            },
+        })
+    }
 }
